+++ spec/sidekiq/transaction_guard/rspec_spec.rb
@@ -3,28 +3,14 @@
 require "spec_helper"
 
 RSpec.describe "rspec integration" do
-  context "with transactional tests" do
-    def use_transactional_tests
-      true
-    end
-
-    it "ignores the outermost transaction" do
-      ActiveRecord::Base.transaction do
-        expect(Sidekiq::TransactionGuard.in_transaction?).to eq false
-      end
+  it "detects transactions in test code" do
+    ActiveRecord::Base.transaction do
+      expect(Sidekiq::TransactionGuard.in_transaction?).to eq true
     end
   end
 
-  context "without transactional tests" do
-    def use_transactional_tests
-      false
-    end
-
-    it "detects the outermost transaction" do
-      ActiveRecord::Base.transaction do
-        expect(Sidekiq::TransactionGuard.in_transaction?).to eq true
-      end
-    end
+  it "is not in a transaction by default" do
+    expect(Sidekiq::TransactionGuard.in_transaction?).to eq false
   end
 
   it "can change the mode for examples", sidekiq_transaction_guard: :stderr do